requirements.txt
numpy
pandas
opencv-python
mediapipe
scikit-learn
matplotlib
pyyaml
pytest
fastapi
uvicorn
python-multipart
streamlit
requests
